chore: refactor zt-plugin for plugin hosting request suggested changes
