+++ .github/CODEOWNERS
@@ -1,0 +1,1 @@
+* @jenkinsci/zerothreat-dast-scanner-plugin-developers
