+++ src/main/java/io/jenkins/plugins/zerothreatai/VulnerabilityScanBuilder.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins;
+package io.jenkins.plugins.zerothreatai;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import hudson.*;
@@ -9,19 +9,20 @@ import hudson.model.TaskListener;
 import hudson.tasks.BuildStepDescriptor;
 import hudson.tasks.BuildStepMonitor;
 import hudson.tasks.Builder;
-import hudson.util.FormValidation;
-import io.jenkins.plugins.models.ScanResponse;
-import io.jenkins.plugins.services.ScanService;
+import hudson.util.Secret;
+import io.jenkins.plugins.zerothreatai.models.ScanResponse;
+import io.jenkins.plugins.zerothreatai.services.ScanService;
 import jenkins.tasks.SimpleBuildStep;
+
+import org.jenkinsci.Symbol;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
-import org.kohsuke.stapler.QueryParameter;
 
 import java.io.IOException;
 import java.io.PrintStream;
 
 public class VulnerabilityScanBuilder extends Builder implements SimpleBuildStep {
-    private String token;
+    private Secret token;
     private boolean waitForAnalysis;
 
     /**
@@ -30,17 +31,17 @@ public class VulnerabilityScanBuilder extends Builder implements SimpleBuildStep
      */
     @DataBoundConstructor
     public VulnerabilityScanBuilder(String token) {
-        this.token = token;
+        this.token = Secret.fromString(token);
         this.waitForAnalysis = false;
     }
 
     public String getToken() {
-        return this.token;
+        return Secret.toString(this.token);
     }
 
     @DataBoundSetter
     public void setToken(String ZT_TOKEN) {
-        this.token = ZT_TOKEN;
+        this.token = Secret.fromString(ZT_TOKEN);
     }
 
     public boolean getWaitForAnalysis() {
@@ -53,9 +54,10 @@ public class VulnerabilityScanBuilder extends Builder implements SimpleBuildStep
     }
 
     @Override
-    public void perform(@NonNull Run<?, ?> run, @NonNull FilePath workspace, @NonNull EnvVars env, @NonNull Launcher launcher, @NonNull TaskListener listener) throws InterruptedException, IOException {
+    public void perform(@NonNull Run<?, ?> run, @NonNull FilePath workspace, @NonNull EnvVars env,
+            @NonNull Launcher launcher, @NonNull TaskListener listener) throws InterruptedException, IOException {
         PrintStream logger = listener.getLogger();
-        if (this.token == null || this.token.trim().isEmpty()) {
+        if (this.getToken() == null || this.getToken().trim().isEmpty()) {
             run.setResult(Result.FAILURE);
             throw new AbortException("Please configure a valid ZT_TOKEN.");
         }
@@ -63,17 +65,17 @@ public class VulnerabilityScanBuilder extends Builder implements SimpleBuildStep
         try {
             // Initiate web scan
             logger.println("Initiating security scan request.");
-            ScanResponse scanResponse = ScanService.initiateScan(this.token, logger);
+            ScanResponse scanResponse = ScanService.initiateScan(this.getToken());
             if (scanResponse.Status != 200) {
                 run.setResult(Result.FAILURE);
-                throw new AbortException("Failed to initiate scan. \nReason: " + scanResponse.Message);
+                throw new AbortException("\nFailed to initiate scan. \nReason: " + scanResponse.Message);
             }
 
-            logger.println("Scan started successfully.\nScan Report Url:" + scanResponse.Url);
+            logger.println("\nScan started successfully.\nScan Report Url:" + scanResponse.Url);
 
             // Poll for scan completion
             if (this.waitForAnalysis) {
-                boolean scanCompleted = ScanService.pollScanStatus(scanResponse.Code, logger);
+                boolean scanCompleted = ScanService.pollScanStatus(this.getToken(), scanResponse.Code, logger);
                 if (!scanCompleted) {
                     run.setResult(Result.FAILURE);
                     throw new AbortException("Scan did not complete successfully.");
@@ -88,8 +90,10 @@ public class VulnerabilityScanBuilder extends Builder implements SimpleBuildStep
     }
 
     @Extension
+    @Symbol("zeroThreatAIScan")
     public static final class DescriptorImpl extends BuildStepDescriptor<Builder> {
         @Override
+        @SuppressWarnings("rawtypes")
         public boolean isApplicable(Class<? extends AbstractProject> jobType) {
             return true;
         }
@@ -98,13 +102,6 @@ public class VulnerabilityScanBuilder extends Builder implements SimpleBuildStep
         @Override
         public String getDisplayName() {
             return "ZeroThreat AI Security Scan";
-        }
-
-        public FormValidation doCheckToken(@QueryParameter String token) {
-            if (token.isBlank()) {
-                return FormValidation.error("ZT_TOKEN is require for initiating scan.");
-            }
-            return FormValidation.ok();
         }
     }
