+++ src/main/java/io/jenkins/plugins/zerothreatai/models/ScanResponse.java
@@ -1,7 +1,4 @@
-package io.jenkins.plugins.models;
-
-
-import javax.annotation.Nullable;
+package io.jenkins.plugins.zerothreatai.models;
 
 public class ScanResponse {
     public int Status = 0;
