+++ src/main/resources/io/jenkins/plugins/zerothreatai/VulnerabilityScanBuilder/config.jelly
@@ -5,7 +5,7 @@
 
         <!-- Required text input for the API token -->
         <f:entry title="ZT_TOKEN" field="token" description="Enter the valid ZT_TOKEN corresponds to target.">
-            <f:textbox/>
+            <f:password clazz="required"/>
         </f:entry>
 
         <!-- Optional boolean checkbox -->
